Add s3

+++ backend/routes/api/spots.js
@@ -5,6 +5,12 @@ const asyncHandler = require("express-async-handler");
 const { Spot, User, Booking } = require("../../db/models");
 const { handleValidationErrors } = require("../../utils/validation");
 const { requireAuth, restoreUser } = require("../../utils/auth");
+const {
+  singleMulterUpload,
+  singlePublicFileUpload,
+  retrievePrivateFile,
+  deleteFile,
+} = require("../../utils/awsS3");
 
 const spotValidation = [
   check("address")
@@ -40,18 +46,19 @@ const spotValidation = [
   check("price")
     .exists({ checkFalsy: true })
     .withMessage("Please provide price."),
-  check("img_link")
-    .exists({ checkFalsy: true })
-    .isLength({ min: 3, max: 255 })
-    .withMessage(
-      "Please provide image link with at least 3 character or less than 255 characters."
-    ),
+  // check("img_link")
+  //   .exists({ checkFalsy: true })
+  //   .isLength({ min: 3, max: 255 })
+  //   .withMessage(
+  //     "Please provide image link with at least 3 character or less than 255 characters."
+  //   ),
   handleValidationErrors,
 ];
 
 router.post(
   "/",
   requireAuth,
+  singleMulterUpload("img_link"),
   spotValidation,
   restoreUser,
   asyncHandler(async (req, res) => {
@@ -59,6 +66,9 @@ router.post(
 
     const { address, city, state, country, lat, lng, name, price, img_link } =
       req.body;
+
+    const spotImageUrl = await singlePublicFileUpload(req.file);
+
     const newSpot = await Spot.create({
       address,
       city,
@@ -69,7 +79,7 @@ router.post(
       name,
       price,
       user_id: user.id,
-      img_link,
+      img_link: spotImageUrl,
     });
     const spot = await Spot.findByPk(newSpot.id, { include: User });
 
@@ -91,21 +101,64 @@ router.get(
 router.delete(
   "/:id",
   asyncHandler(async (req, res) => {
+    const spot = await Spot.findByPk(req.params.id);
+    if (!spot) throw new Error("No Campsite with that id");
+
     await Booking.destroy({
       where: { spot_id: req.params.id },
     });
-    const deletedSpot = await Spot.deleteSpot(req.params.id);
-    return res.json(deletedSpot);
+    // const deletedSpot = await Spot.deleteSpot(req.params.id);
+    await deleteFile(spot.img_link.split("com/")[1]);
+    await spot.destroy();
+
+    return res.json(spot);
   })
 );
 
 router.put(
   "/:id",
   requireAuth,
+  singleMulterUpload("img_link"),
   spotValidation,
   asyncHandler(async (req, res) => {
-    const id = await Spot.updateSpot(req.body);
-    const spot = await Spot.getOne(id);
+    // const id = await Spot.updateSpot(req.body);
+
+    const id = req.body.id;
+    delete req.body.id;
+
+    const { address, city, state, country, lat, lng, name, price, user_id } =
+      req.body;
+
+    let spot = await Spot.findByPk(id);
+    console.log(id);
+    console.log("\n\n\n", spot, "\n\n\n");
+
+    let spotImageUrl;
+    if (req.file) {
+      spotImageUrl = await singlePublicFileUpload(req.file);
+      await deleteFile(spot.img_link.split("com/")[1]);
+    } else {
+      spotImageUrl = spot.img_link;
+    }
+
+    await Spot.update(
+      {
+        address,
+        city,
+        state,
+        country,
+        lat,
+        lng,
+        name,
+        price,
+        user_id,
+        img_link: spotImageUrl,
+      },
+      { where: { id }, returning: true, plain: true }
+    );
+
+    spot = await Spot.findByPk(id);
+
     return res.json(spot);
   })
 );
